+++ packages/moonstone/ToggleButton/ToggleButton.js
@@ -10,6 +10,7 @@
  */
 
 import kind from '@enact/core/kind';
+import deprecate from '@enact/core/internal/deprecate';
 import React from 'react';
 import PropTypes from 'prop-types';
 import Pure from '@enact/ui/internal/Pure';
@@ -19,6 +20,19 @@ import Button from '../Button';
 import Skinnable from '../Skinnable';
 
 import css from './ToggleButton.module.less';
+
+const deprecateSmall = deprecate((small) => small ? 'small' : 'large',  {
+	name: 'moonstone/ToggleButton.ToggleButtonBase#small',
+	replacedBy: 'the `size` prop',
+	message: 'Use `size="small"` instead.',
+	since: '2.6.0',
+	until: '3.0.0'
+});
+
+function getSize (size, small) {
+	small = typeof small !== 'undefined' ? deprecateSmall(small) : 'large';
+	return size || small;
+}
 
 /**
  * A stateless [Button]{@link moonstone/Button.Button} that can be toggled by changing its
@@ -94,13 +108,22 @@ const ToggleButtonBase = kind({
 		selected: PropTypes.bool,
 
 		/**
+		 * The size of the button.
+		 *
+		 * A `'small'` button will have a larger tap target than its apparent size to allow it to be
+		 * clicked more easily.
+		 *
+		 * @type {('small'|'large')}
+		 * @default 'large'
+		 * @public
+		 */
+		size: PropTypes.string,
+
+		/**
 		 * Reduces the size of the button.
 		 *
-		 * The button will have a larger tap target than its apparent size to allow it to be clicked
-		 * more easily.
-		 *
 		 * @type {Boolean}
-		 * @default false
+		 * @deprecated replaced by prop `size='small'`
 		 * @public
 		 */
 		small: PropTypes.bool,
@@ -132,7 +155,7 @@ const ToggleButtonBase = kind({
 		disabled: false,
 		minWidth: true,
 		selected: false,
-		small: false,
+		// size: 'large', // we won't set default props for `size` yet to support `small` prop
 		toggleOffLabel: '',
 		toggleOnLabel: ''
 	},
@@ -143,7 +166,7 @@ const ToggleButtonBase = kind({
 	},
 
 	computed: {
-		className: ({selected, small, styler}) => styler.append({selected, small}),
+		className: ({selected, size, small, styler}) => styler.append({selected}, getSize(size, small)),
 		children: ({children, selected, toggleOnLabel, toggleOffLabel}) => {
 			let c = children;
 			if (selected && toggleOnLabel) {
@@ -158,6 +181,7 @@ const ToggleButtonBase = kind({
 	render: ({selected, ...rest}) => {
 		delete rest.toggleOffLabel;
 		delete rest.toggleOnLabel;
+		delete rest.small;
 
 		return (
 			<Button data-webos-voice-intent="SetToggleItem" {...rest} aria-pressed={selected} selected={selected} />