+++ packages/moonstone/TooltipDecorator/TooltipDecorator.js
@@ -116,7 +116,7 @@ const TooltipDecorator = hoc(defaultConfig, (config, Wrapped) => {
 			 * | `'above right'` | Above component, flowing to the right |
 			 * | `'below'` | Below component, flowing to the right |
 			 * | `'below center'` | Below component, centered |
-			 * | `'below left'` | Below component, flowing to the right |
+			 * | `'below left'` | Below component, flowing to the left |
 			 * | `'below right'` | Below component, flowing to the right |
 			 * | `'left bottom'` | Left of the component, contents at the bottom |
 			 * | `'left middle'` | Left of the component, contents middle aligned |